(un-tested) conversion from ASSIMP to desired scene/mesh/material format.
Need to add in creation of 'scene' data (such that a scene exists in the database and holds references to each mesh it uses, and that each mesh exists in the database and holds references to each material they use, and each material exists in the database. Each scene, mesh, and material should also contain the file that holds their information in the database.

Though in the case of materials, we could change this to not have a file for them, and store their information directly in the databse.